# docs/swagger.yml
openapi: 3.0.0
info:
  title: SwapTime Imager
  description: Imager server that stores images by categories.
  version: 1.0.0
paths:
  /api/v1/categories:
    get:
      tags: [public]
      summary: Returns a list of categories.
      responses:
        "200":
          description: List of categories.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "500":
          description: Internal server error.
        "503":
          description: Service unavailable.
  /api/v1/images:
    get:
      tags: [public]
      summary: List images.
      parameters:
      - name: limit
        in: query
        schema:
          type: number
          minimum: 1
        required: true
      - name: offset
        in: query
        schema:
          type: number
          minimum: 0
        required: true
      - name: category
        in: query
        schema:
          type: string
        required: true
      responses:
        "200":
          description: Images.
          content:
            "application/json":
              schema: 
                type: array
                items:
                  $ref: "#/components/schemas/ImageMeta"
        "400":
          description: Bad request.
        "500":
          description: Internal server error.
        "503":
          description: Service unavailable.
  /api/v1/images/{id}/{size}:
    get:
      tags: [public]
      summary: Get image data.
      parameters:
      - name: id
        in: path
        schema:
          type: string
          format: uuid
        required: true
      - name: size
        in: path
        schema:
          type: string
          example: 1080x1920
        required: true
      responses:
        "200":
          description: Image body.
          content:
            "image/*":
              schema: 
                type: string
                format: binary
        "500":
          description: Internal server error.
        "503":
          description: Service unavailable.
  /internal/api/v1/images/shuffle:
    post:
      tags: [internal]
      summary: Shuffle images in the category.
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              required:
              - category
              - depth
              properties:
                category:
                  type: string
                  example: all
                depth:
                  type: integer
                  example: 512
      responses:
        "200":
          description: Image metadata.
          content:
            application/json:
              schema:
                type: string
                example: ok
        "400":
          description: Bad request.
        "500":
          description: Internal server error.
        "503":
          description: Service unavailable.
  # TODO: delete method.
  /internal/api/v1/images:
    put:
      tags: [internal]
      summary: Upload an image.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - author
              - websource
              - category
              - image
              properties:
                id:
                  type: string
                  format: uuid
                author:
                  type: string
                websource:
                  type: string
                category:
                  type: string
                  description: Category should not be "all". This name is reserved.
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: Image metadata.
          content:
            application/json:
              schema:
                $ref:  "#/components/schemas/ImageMeta"
        "400":
          description: Bad request.
        "409":
          description: Conflict.
        "413":
          description: Request entity too large.
        "415":
          description: Unsupported media type.
        "500":
          description: Internal server error.
        "503":
          description: Service unavailable.
components:
  schemas:
    ImageMeta:
      type: object
      required:
      - id
      - author
      - source
      - mimetype
      - category
      properties:
        id:
          type: string
          format: uuid
        author:
          type: string
        source:
          type: string
        mimetype:
          type: string
        category:
          type: string